API version check: requirements & tracing
