+++ common/src/helpers.rs
@@ -14,6 +14,9 @@
 use serde::{Serialize, Serializer};
 use std::collections::{BTreeMap, HashMap};
 
+use crate::{std_extensions::IllegalStateResult, ANKAIOS_VERSION};
+use semver::Version;
+
 // [impl->swdd~common-helper-methods~1]
 pub fn try_into_vec<S, T, E>(input: Vec<S>) -> Result<Vec<T>, E>
 where
@@ -31,4 +34,74 @@ where
 {
     let ordered: BTreeMap<_, _> = value.iter().collect();
     ordered.serialize(serializer)
+}
+
+pub fn check_version_compatibility(version: impl AsRef<str>) -> Result<(), String> {
+    let ank_version = Version::parse(ANKAIOS_VERSION).unwrap_or_illegal_state();
+    if let Ok(input_version) = Version::parse(version.as_ref()) {
+        if ank_version.major == input_version.major &&
+        // As we are at a 0 (zero) major version, we also require minor version equality
+        ank_version.minor == input_version.minor
+        {
+            return Ok(());
+        }
+    } else {
+        log::warn!(
+            "Could not parse incoming string '{}' as semantic version.",
+            version.as_ref()
+        );
+    };
+
+    Err(format!(
+        "Unsupported protocol version '{}'. Currently supported '{ANKAIOS_VERSION}'",
+        version.as_ref()
+    ))
+}
+
+//////////////////////////////////////////////////////////////////////////////
+//                 ########  #######    #########  #########                //
+//                    ##     ##        ##             ##                    //
+//                    ##     #####     #########      ##                    //
+//                    ##     ##                ##     ##                    //
+//                    ##     #######   #########      ##                    //
+//////////////////////////////////////////////////////////////////////////////
+
+#[cfg(test)]
+mod tests {
+    use semver::Version;
+
+    use crate::{check_version_compatibility, ANKAIOS_VERSION};
+
+    // [utest->swdd~common-version-checking~1]
+    #[test]
+    fn utest_version_compatibility_success() {
+        assert!(check_version_compatibility(ANKAIOS_VERSION).is_ok())
+    }
+
+    // [utest->swdd~common-version-checking~1]
+    #[test]
+    fn utest_version_compatibility_patch_diff_success() {
+        let mut version = Version::parse(ANKAIOS_VERSION).unwrap();
+        version.patch = 199;
+        assert!(check_version_compatibility(version.to_string()).is_ok())
+    }
+
+    // [utest->swdd~common-version-checking~1]
+    #[test]
+    fn utest_version_compatibility_patch_major_error() {
+        let mut version = Version::parse(ANKAIOS_VERSION).unwrap();
+        version.major = 199;
+        assert!(check_version_compatibility(version.to_string()).is_err())
+    }
+
+    // [utest->swdd~common-version-checking~1]
+    #[test]
+    fn utest_version_compatibility_patch_minor_error() {
+        let mut version = Version::parse(ANKAIOS_VERSION).unwrap();
+        version.minor = 199;
+        // Currently we assert that the minor version is also equal as we are at a 0th major version.
+        // When a major version is released, we can update the test here and expect an Ok().
+        assert_eq!(0, version.major);
+        assert!(check_version_compatibility(version.to_string()).is_err())
+    }
 }
